php-blog-mysql
index.php is root file. Click on create post, type a title and a description and then click on add post.  Click on Read Me button to view content. Click on Edit button and change title then click on Update to update. And finally to click on Delete to delete post